+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## 1.31
 _WIP_
 
+**Backend**
+- Do not re-fetch submissions on terminal screen resize events when screen height 
+remains the same
+
 **Cosmetic**
 - Updated filter list for macrumors.com